Landing page

+++ backend/migrations/20191024161823-reviews.js
@@ -9,7 +9,7 @@ module.exports = {
                 autoIncrement: true
             },
             rating: {
-                type: Sequelize.INTEGER(1),
+                type: Sequelize.SMALLINT,
                 allowNull: false
             },
             userId: {
